feat: Add Tag CT
