+++ src/docs/Tag.mdx
@@ -99,6 +99,9 @@ function Component() {
     <Tag size="medium" theme="link">
       Tag Link
     </Tag>
+    <Tag size="medium" theme="promo">
+      Tag Promo
+    </Tag>
   </Flexbox>
 </Demo>
